Remove TryRemovingNullCheck from LSE
TryRemovingNullCheck would remove null checks right before eliminated loads. However, it was doing it without really checking if the object it was loading from was null or not.

This has been there since the first iteration of LSE, and my guess why it was implemented it was to remove NullChecks with no uses (i.e. it was assuming that the NullCheck was only uses by the load which was being eliminated).

Bug: 243136201
Test: Ran ART benchmark automator and saw no increase in code size
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Change-Id: I96be9650f8b16fee7fcca114b9b330d4f3b7e0b4
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: tejasvp25 <tejasvp25@gmail.com>