Don't disable JIT GC when enabling method tracing
Earlier we used to disable jit GC when enabling method tracing. This was
because instrumentation stubs fetched jit code for the method if it
existed. There could be races where the code was fetched but jit GC
cleared it before the jit code could be invoked (b/110263880). Currently,
JITed code has instrumentation inbuilt and we don't fetch jited code
from instrumentation stub. So we no longer need to disable jit GC.

Bug: 206029744
Bug: 110263880
Test: art/testrunner.py --trace
Change-Id: Ia5b7daa9d05e43add9486abf327dc4390c09ef53
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: tejasvp25 <tejasvp25@gmail.com>