Improve linear loop optimization overflow checks
We can move the logic from GenerateLastValueLinear to GenerateCode,
where we can:
 * be more granular, and
 * reuse that code for other methods.

To do so we add an allow_potential_overflow variable. This is done
because BCE does perform operations that might overflow, but it
uses HDeoptimize instructions to guard against that. Loop
optimization doesn't add HDeoptimize instructions so we must be
more careful.

Bug: 231415860
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Change-Id: I1d689e5ddd87d96707673b6065ab0cc48b288617
Signed-off-by: Dmitrii <bankersenator@gmail.com>